display OSD icons other than play for osd_level=1
A tweak to osd_level=1. Causes seek, pause, and buffering to display the
associated icon in the top left OSD position.

This a significant UI change, but fits the spirit of osd_level=1 (OSD is
displayed for user input, not otherwise) and resembles the UI of other
video and players.